Делаю рейсы.Нужно сделать валидацию и временную таблицу

+++ main/java/com/svop/tables/Handbooks/NomerReysRepository.java
@@ -5,11 +5,13 @@ import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 
-public interface NomerReysRepository extends JpaRepository<NomerReys,Long> {
+public interface NomerReysRepository extends JpaRepository<NomerReys,Integer> {
     List<NomerReys> findByAircompany_Id(Integer id);
     @Transactional
     void deleteByIdIn(Iterable<Integer> id);
     @Transactional
     void save(Iterable<NomerReys> nomerReys);
+
+
 
 }
