+++ main/java/com/svop/tables/Handbooks/RoutesRepository.java
@@ -6,7 +6,7 @@ import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 
-public interface RoutesRepository extends JpaRepository<Routes,Long> {
+public interface RoutesRepository extends JpaRepository<Routes,Integer> {
     @Transactional
     void deleteByIdIn(List<Integer> ids);
 }
